back-end/requirements.txt
Django==4.1.6
djangorestframework==3.11.2
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
drf-spectacular>=0.15.1, <0.16
django-filter==22.1
pytz==2022.7.1